refactor: 패키지명, 클래스명 수정
