+++ src/main/kotlin/racingcar/dto/RacingResultDTO.kt
@@ -1,3 +1,3 @@
-package racingcar.etc
+package racingcar.dto
 
 data class RacingResultDTO(val distance: Int)
