+++ src/main/kotlin/racingcar/ui/GameManager.kt
@@ -1,4 +1,4 @@
-package racingcar.etc
+package racingcar.ui
 
 import racingcar.domain.Cars
